requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
distro==1.5.0
idna==2.10
lxml==4.5.1
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tabula-py==2.1.1
urllib3==1.25.9
wget==3.2
